shorten snippet name for USBMIDI
