+++ APIs_USB/USBMIDI_Take_Me_Out/README.md
@@ -1,0 +1,9 @@
+## USBMIDI "Take Me Out to the Ball Game" example
+
+The example plays an entire song, not just a series of notes. "Take Me Out to the Ball Game" is a popular song in the public domain.
+
+1. Flash the board, and ensure the target's auxiliary USB is plugged into the PC.
+2. Open Anvil Studio.
+3. Click View > Synthesizers, MIDI + Audio Devices.
+4. Uncheck Synth is too slow to echo incoming events.
+5. Click View > Composer (Staff Editor) to see notes from the board being mapped to the sheet music.
